move UserRepository package
