+++ app/src/main/java/com/wafflestudio/snugo/di/RepositoryModule.kt
@@ -1,7 +1,7 @@
 package com.wafflestudio.snugo.di
 
-import com.wafflestudio.snugo.features.onboarding.UserRepository
-import com.wafflestudio.snugo.features.onboarding.UserRepositoryImpl
+import com.wafflestudio.snugo.repository.UserRepository
+import com.wafflestudio.snugo.repository.UserRepositoryImpl
 import com.wafflestudio.snugo.repository.BuildingsRepository
 import com.wafflestudio.snugo.repository.BuildingsRepositoryImpl
 import dagger.Binds
