+++ app/src/main/java/com/wafflestudio/snugo/features/onboarding/UserViewModel.kt
@@ -2,6 +2,7 @@ package com.wafflestudio.snugo.features.onboarding
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.wafflestudio.snugo.repository.UserRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
